Switch to fabric instead of capistrano dependency.
Simplifies things greatly, since we now don't need to git clone and run cap commands locally, but instead just execute those things on the remote host directly.

Plus, it's all python now, and the lack of clunky ssh/sftp stuff in the main code relieves us of a few module dependencies (or at least they are implied due to fabric's own dependencies)

This still leaves the problem of having to pull on the remote host from a private repo, which I hope to rethink out later.

Introduce Puppet simulate option in the config ini so that --noop can be run.